Added data-class and data-innerHTML
data-innerHTML is the new name for data-content to prepare for version
2. data-content is maintained for legacy support, but the plan is to
remove this in version 2. data-template-bind will also accept
"innerHTML" as a replacement to content. Again the plan is to free up
content, as this could be a valid attribute of some elements. data-class
will bind to the class attribute